requirements.txt
pandas==1.1.5
pyinstaller==4.10.0
python-dateutil==2.8.2